Revert "small code refactoring"
Inadvertently deleted a file.  Reverting
