+++ model/adoptiondata.py
@@ -10,7 +10,8 @@ from model.metaclass_cache import MetaclassCache
 import numpy as np
 import pandas as pd
 
-from model.data_handler import DataHandler, data_func
+from model.data_handler import DataHandler
+from model.decorators import data_func
 
 class AdoptionData(DataHandler, object, metaclass=MetaclassCache):
     """Implements Adoption Data module."""
