+++ model/decorators.py
@@ -1,0 +1,6 @@
+"""
+Set of decorators for model data presentation
+"""
+def data_func(method):
+    method.data_func = True
+    return method
