+++ model/dez.py
@@ -17,7 +17,8 @@ import pandas as pd
 from model import dd
 from model.metaclass_cache import MetaclassCache
 
-from model.data_handler import DataHandler, data_func
+from model.data_handler import DataHandler
+from model.decorators import data_func
 
 OCEAN_CSV_PATH = pathlib.Path(__file__).parents[1].joinpath('data', 'ocean')
